add 2 error test case
