+++ test/msg_delay_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 func genMsgDelay(gid string) *dtmcli.Msg {
-	req := busi.GenTransReq(30, false, false)
+	req := busi.GenReqHTTP(30, false, false)
 	msg := dtmcli.NewMsg(dtmutil.DefaultHTTPServer, gid).
 		Add(busi.Busi+"/TransOut", &req).
 		Add(busi.Busi+"/TransIn", &req).SetDelay(10)
