+++ test/saga_test.go
@@ -78,7 +78,7 @@ func TestSagaAbnormal(t *testing.T) {
 
 func TestSagaEmptyUrl(t *testing.T) {
 	saga := dtmcli.NewSaga(dtmutil.DefaultHTTPServer, dtmimp.GetFuncName())
-	req := busi.GenTransReq(30, false, false)
+	req := busi.GenReqHTTP(30, false, false)
 	saga.Add(busi.Busi+"/TransOut", "", &req)
 	saga.Add("", "", &req)
 	saga.Submit()
@@ -89,7 +89,7 @@ func TestSagaEmptyUrl(t *testing.T) {
 
 func genSaga(gid string, outFailed bool, inFailed bool) *dtmcli.Saga {
 	saga := dtmcli.NewSaga(dtmutil.DefaultHTTPServer, gid)
-	req := busi.GenTransReq(30, outFailed, inFailed)
+	req := busi.GenReqHTTP(30, outFailed, inFailed)
 	saga.Add(busi.Busi+"/TransOut", busi.Busi+"/TransOutRevert", &req)
 	saga.Add(busi.Busi+"/TransIn", busi.Busi+"/TransInRevert", &req)
 	return saga
@@ -97,7 +97,7 @@ func genSaga(gid string, outFailed bool, inFailed bool) *dtmcli.Saga {
 
 func genSaga1(gid string, outFailed bool, inFailed bool) *dtmcli.Saga {
 	saga := dtmcli.NewSaga(dtmutil.DefaultHTTPServer, gid)
-	req := busi.GenTransReq(30, outFailed, inFailed)
+	req := busi.GenReqHTTP(30, outFailed, inFailed)
 	saga.Add(busi.Busi+"/TransOut", busi.Busi+"/TransOutRevert", &req)
 	return saga
 }
